requirements.txt
llama-cloud-services 
llama-index-core 
llama-index-readers-file 
python-dotenv